Everything working commit

+++ src/main/java/com/wang/michael/online_shop/service/NewsService.java
@@ -9,11 +9,11 @@ public interface NewsService {
 
   public News create(News news);
 
-  public News delete(int id) throws NewsNotFound;
+  public News delete(Long id) throws NewsNotFound;
 
   public List<News> findAll();
 
   public News update(News news) throws NewsNotFound;
 
-  public News findById(int id) throws NewsNotFound;
+  public News findById(Long id) throws NewsNotFound;
 }
